resolve #44 "process_time" is different from the actual processing time

+++ src/PhpJsonLogger/Logger.php
@@ -202,9 +202,9 @@ class Logger
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getCreatedTime(): int
+    public function getCreatedTime(): float
     {
         return $this->createdTime;
     }
